Module leveldb::database

The main database module, allowing to interface with leveldb on a key-value basis.

Modules

batch
Module providing write batches

cache
Structs and traits to work with the leveldb cache.

comparator
All keys in leveldb are compared by their binary value unless defined otherwise.

error
The module defining custom leveldb error type.

iterator
leveldb iterators

kv
Key-Value semantics.

options
All the option types needed for interfacing with leveldb.

snapshots
leveldb snapshots

Structs

Database
The main database object.