Module leveldb::database::comparator

All keys in leveldb are compared by their binary value unless defined otherwise.

Comparators allow to override this comparison. The ordering of keys introduced by the compartor influences iteration order. Databases written with one Comparator cannot be opened with another.

Structs

DefaultComparator
DefaultComparator is the a stand in for "no comparator set"

OrdComparator
OrdComparator is a comparator comparing Keys that implement Ord

Traits

Comparator
A comparator has two important functions:

Functions

create_comparator