Module leveldb::database::iterator

leveldb iterators

Iteration is one of the most important parts of leveldb. This module provides Iterators to iterate over key, values and pairs of both.

Structs

Iterator
An iterator over the leveldb keyspace.

KeyIterator
An iterator over the leveldb keyspace.

ValueIterator
An iterator over the leveldb keyspace.

Traits

Iterable
A trait to allow access to the three main iteration styles of leveldb.

LevelDBIterator

LevelDBIteratorExt