Module leveldb::database::options

All the option types needed for interfacing with leveldb.

Those are: * Options: used when opening a database * ReadOptions: used when reading from leveldb * WriteOptions: used when writng to leveldb

Structs

Options
Options to consider when opening a new or pre-existing database.

ReadOptions
The read options to use for any read operation.

WriteOptions
The write options to use for a write operation.

Functions

c_options

c_readoptions

c_writeoptions