Struct leveldb::database::options::Options

pub struct Options {
    pub create_if_missing: bool,
    pub error_if_exists: bool,
    pub paranoid_checks: bool,
    pub write_buffer_size: Option<size_t>,
    pub max_open_files: Option<i32>,
    pub block_size: Option<size_t>,
    pub block_restart_interval: Option<i32>,
    pub compression: Compression,
    pub cache: Option<Cache>,
}

Options to consider when opening a new or pre-existing database.

Note that in contrast to the leveldb C API, the Comparator is not passed using this structure.

For more detailed explanations, consider the leveldb documentation

Fields

create_if_missing

create the database if missing

default: false

error_if_exists

report an error if the DB already exists instead of opening.

default: false

paranoid_checks

paranoid checks make the database report an error as soon as corruption is detected.

default: false

write_buffer_size

Override the size of the write buffer to use.

default: None

max_open_files

Override the max number of open files.

default: None

block_size

Override the size of the blocks leveldb uses for writing and caching.

default: None

block_restart_interval

Override the interval between restart points.

default: None

compression

Define whether leveldb should write compressed or not.

default: Compression::No

cache

A cache to use during read operations.

default: None

Methods

impl Options

fn new() -> Options

Create a new Options struct with default settings.