Struct leveldb::database::options::ReadOptions

pub struct ReadOptions<'a, K: Key + 'a> {
    pub verify_checksums: bool,
    pub fill_cache: bool,
    pub snapshot: Option<&'a Snapshot<'a, K>>,
}

The read options to use for any read operation.

Fields

verify_checksums

Whether to verify the saved checksums on read.

default: false

fill_cache

Whether to fill the internal cache with the results of the read.

default: true

snapshot

An optional snapshot to base this operation on.

Consider using the Snapshot trait instead of setting this yourself.

default: None

Methods

impl<'a, K: Key + 'a> ReadOptions<'a, K>

fn new() -> ReadOptions<'a, K>

Return a ReadOptions struct with the default values.