Struct leveldb::database::Database

pub struct Database<K: Key> {
    // some fields omitted
}

The main database object.

leveldb databases are based on ordered keys. By default, leveldb orders by the binary value of the key. Additionally, a custom Comparator can be passed when opening the database. This library ships with an Comparator implementation for keys that are Ord.

When re-opening a database, you must use the same key type K and comparator type C.

Multiple Database objects can be kept around, as leveldb synchronises internally.

Methods

impl<K: Key> Database<K>

fn open(name: &Path, options: Options) -> Result<Database<K>, Error>

Open a new database

If the database is missing, the behaviour depends on options.create_if_missing. The database will be created using the settings given in options.

fn open_with_comparator<C: Comparator<K=K>>(name: &Path, options: Options, comparator: C) -> Result<Database<K>, Error>

Open a new database with a custom comparator

If the database is missing, the behaviour depends on options.create_if_missing. The database will be created using the settings given in options.

The comparator must implement a total ordering over the keyspace.

For keys that implement Ord, consider the OrdComparator.